requirements.txt
django==4.2.2
pillow==9.5.0
tensorflow
numpy
opencv-python
weasyprint
mysqlclient
djangorestframework==3.14.0
drf-yasg==1.21.6